requirements.txt
alembic~=1.13.1
asyncpg==0.29.0
Authlib==1.3.0
bcrypt==4.1.3
fastapi~=0.110.3
fastapi-injector==0.5.4
fastapi-jwt==0.3.0
injector~=0.21.0
loguru==0.6.0
logtail-python==0.3.0
mediatr~=1.3.2
SQLAlchemy~=2.0.29
uvicorn==0.29.0
pydantic~=2.7.1
pyjwt==2.3.0
timestamp~=0.0.1
